+++ src/config/shopConfig.js
@@ -79,8 +79,7 @@ const shopItemConfig = {
         "description": "Gain a random Pokeball used to catch Pokemon! Use the \`/help gacha\` command to learn more! Limit 5 per day.",
         "category": shopCategories.GACHA,
         "price": [200],
-        // TEMP: gen 2 event
-        "limit": 5 * 2,
+        "limit": 5,
     },
     [shopItems.HOME]: {
         "name": "Home",